Added toString() method to EventSeverity for clarity in debugger.
Altered the way 'events' are passed through claims generation. Currently we extract severities and dates at the top of the stack, then pass down two lists. It is imperitive that the lists remain in the same order. Instead of relying the lists not changing (which is not guaranteed by the code, only by developer convention) as they are passed down the stack, instead do the extraction of the Event / Severity at the bottom of the stack at the moment of claims generation. The order of extraction is guaranteed (and we save about 6 lines of code :-)...